requirements.txt
torch>=2.0.0
transformers>=4.57.0
Pillow>=9.0.0
gradio>=4.0.0
qwen-vl-utils
accelerate>=0.26.0
